requirements.txt
django>=5.0
djangorestframework>=3.15
django-filter>=24.0
